requirements.txt
pip==21.3.1
requests>=2.27.1
selenium>=3.141.0
pytest>=6.2.5
ipython>=5.10.0
importlib-metadata>=2.1.1
virtualenv>=20.13.0
jsonpickle>=2.1.0
psycopg>=3.0.7